move layout to template
