+++ theme/gatsby-node.js
@@ -1,7 +1,7 @@
 exports.createPages = ({ actions }) => {
   actions.createPage({
     path: "/",
-    component: require.resolve("./src/templates/page.js"),
+    component: require.resolve("./src/templates/page.jsx"),
     context: {
       heading: "Scarlet",
       content: `<p>test</p>`,
