+++ theme/src/gatsby-plugin-theme-ui/index.js
@@ -55,7 +55,6 @@ export default {
     },
     // Theme UI Components
     Layout: {
-      // backgroundColor: "background",
       color: "text",
       fontFamily: "body",
       fontSize: 1,
