--- theme/src/templates/projects-listing.js
@@ -1,1 +1,0 @@
-// TODO
