--- theme/src/templates/projects-single.js
@@ -1,1 +1,0 @@
-// TODO
